Use inhibit-read-only instead of toggle-read-only, which is obsolete
Fixing byte-compile warnings about:

  'toggle-read-only' is an obsolete function (as of 24.3);
  use 'read-only-mode' instead.

Note, we're not using read-only-mode because its docstring says:

  "Do not call this from a Lisp program unless you really intend to
  do the same thing as the C-x C-q command, including
  possibly enabling or disabling View mode."

Also, a commit message by Philip Weaver said:

  "setting inhibit-read-only is a bad idea.  it globally affects all
  buffers and basically makes it so read-only is completely ignored.
  it's better to turn off read-only mode locally, make the changes,
  then turn it back on again."

In fact, it is a bad idea only when one does a

  (setq inhibit-read-only t)

as the original code that Philip Weaver patched did.  There is
no problem when you set it in a let, as this commit does.

Also, using lexical bindings should further mitigate this problem.